# frontend/src/generated/openapi.json
{
  "components": {
    "schemas": {
      "ABEntryResponse": {
        "additionalProperties": false,
        "properties": {
          "config_version": {
            "title": "Config Version",
            "type": "string"
          },
          "evaluated_target_count": {
            "title": "Evaluated Target Count",
            "type": "integer"
          },
          "scores": {
            "additionalProperties": {
              "$ref": "#/components/schemas/MeanCountResponse"
            },
            "title": "Scores",
            "type": "object"
          }
        },
        "required": [
          "config_version",
          "evaluated_target_count",
          "scores"
        ],
        "title": "ABEntryResponse",
        "type": "object"
      },
      "AnalysisResponse": {
        "additionalProperties": false,
        "properties": {
          "ab_leaderboard": {
            "items": {
              "$ref": "#/components/schemas/ABEntryResponse"
            },
            "title": "Ab Leaderboard",
            "type": "array"
          },
          "coaching_markdown": {
            "title": "Coaching Markdown",
            "type": "string"
          },
          "summary": {
            "items": {
              "$ref": "#/components/schemas/ScoreSummaryResponse"
            },
            "title": "Summary",
            "type": "array"
          },
          "trends": {
            "items": {
              "$ref": "#/components/schemas/TrendEntryResponse"
            },
            "title": "Trends",
            "type": "array"
          }
        },
        "required": [
          "summary",
          "ab_leaderboard",
          "trends",
          "coaching_markdown"
        ],
        "title": "AnalysisResponse",
        "type": "object"
      },
...
      "FeedbackItemResponse": {
        "additionalProperties": true,
        "properties": {
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "feedback_type": {
            "title": "Feedback Type",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "payload": {
            "$ref": "#/components/schemas/FeedbackPayloadResponse"
          }
        },
        "required": [
          "id",
          "feedback_type",
          "payload",
          "created_at"
        ],
        "title": "FeedbackItemResponse",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "JudgingTokenEstimateResponse": {
        "additionalProperties": false,
        "properties": {
          "largest_turn_tokens": {
            "title": "Largest Turn Tokens",
            "type": "integer"
          },
          "total_tokens": {
            "title": "Total Tokens",
            "type": "integer"
          },
          "turn_tokens": {
            "items": {
              "type": "integer"
            },
            "title": "Turn Tokens",
            "type": "array"
          }
        },
        "required": [
          "total_tokens",
          "largest_turn_tokens",
          "turn_tokens"
        ],
        "title": "JudgingTokenEstimateResponse",
        "type": "object"
      },
...
      "ScoreSummaryResponse": {
        "additionalProperties": false,
        "properties": {
          "binary": {
            "title": "Binary",
            "type": "boolean"
          },
          "ci": {
            "maxItems": 2,
            "minItems": 2,
            "prefixItems": [
              {
                "type": "number"
              },
              {
                "type": "number"
              }
            ],
            "title": "Ci",
            "type": "array"
          },
          "confident": {
            "title": "Confident",
            "type": "boolean"
          },
          "count": {
            "title": "Count",
            "type": "integer"
          },
          "mean": {
            "title": "Mean",
            "type": "number"
          },
          "pass_rate": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Pass Rate"
          },
          "scorer": {
            "title": "Scorer",
            "type": "string"
          },
          "tag_counts": {
            "additionalProperties": {
              "type": "integer"
            },
            "title": "Tag Counts",
            "type": "object"
          }
        },
        "required": [
          "scorer",
          "count",
          "mean",
          "ci",
          "binary",
          "pass_rate",
          "tag_counts",
          "confident"
        ],
        "title": "ScoreSummaryResponse",
        "type": "object"
      },
...
      "SessionDetailResponse": {
        "additionalProperties": false,
        "properties": {
          "config_version": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Config Version"
          },
          "conversation_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Conversation Id"
          },
          "git_branch": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Git Branch"
          },
          "judging_token_estimates": {
            "additionalProperties": {
              "$ref": "#/components/schemas/JudgingTokenEstimateResponse"
            },
            "propertyNames": {
              "enum": [
                "utf8_bytes_div_3",
                "o200k_base",
                "o200k_harmony"
              ]
            },
            "title": "Judging Token Estimates",
            "type": "object"
          },
          "session_feedback": {
            "items": {
              "$ref": "#/components/schemas/FeedbackItemResponse"
            },
            "title": "Session Feedback",
            "type": "array"
          },
          "signal_evidence": {
            "items": {
              "$ref": "#/components/schemas/SignalEvidenceResponse"
            },
            "title": "Signal Evidence",
            "type": "array"
          },
          "total_tokens": {
            "title": "Total Tokens",
            "type": "integer"
          },
          "turn_count": {
            "title": "Turn Count",
            "type": "integer"
          },
          "turn_feedback": {
            "additionalProperties": {
              "items": {
                "$ref": "#/components/schemas/FeedbackItemResponse"
              },
              "type": "array"
            },
            "title": "Turn Feedback",
            "type": "object"
          },
          "turns": {
            "items": {
              "$ref": "#/components/schemas/TurnDetailResponse"
            },
            "title": "Turns",
            "type": "array"
          }
        },
        "required": [
          "conversation_id",
          "config_version",
          "git_branch",
          "total_tokens",
          "turn_count",
          "judging_token_estimates",
          "turns",
          "session_feedback",
          "turn_feedback",
          "signal_evidence"
        ],
        "title": "SessionDetailResponse",
        "type": "object"
      },
      "SessionListResponse": {
        "additionalProperties": false,
        "properties": {
          "limit": {
            "title": "Limit",
            "type": "integer"
          },
          "sessions": {
            "items": {
              "$ref": "#/components/schemas/SessionSummaryResponse"
            },
            "title": "Sessions",
            "type": "array"
          },
          "total": {
            "title": "Total",
            "type": "integer"
          },
          "truncated": {
            "title": "Truncated",
            "type": "boolean"
          }
        },
        "required": [
          "sessions",
          "total",
          "truncated",
          "limit"
        ],
        "title": "SessionListResponse",
        "type": "object"
      },
      "SessionSummaryResponse": {
        "additionalProperties": false,
        "properties": {
          "config_version": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Config Version"
          },
          "conversation_id": {
            "title": "Conversation Id",
            "type": "string"
          },
          "effort_level": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Effort Level"
          },
          "ended_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ended At"
          },
          "git_branch": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Git Branch"
          },
          "input_preview": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Input Preview"
          },
          "last_activity": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Activity"
          },
          "model": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Model"
          },
          "session_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Session Id"
          },
          "signal_evidence": {
            "items": {
              "$ref": "#/components/schemas/SignalEvidenceResponse"
            },
            "title": "Signal Evidence",
            "type": "array"
          },
          "started_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Started At"
          },
          "total_tokens": {
            "title": "Total Tokens",
            "type": "integer"
          },
          "total_tool_calls": {
            "title": "Total Tool Calls",
            "type": "integer"
          },
          "turn_count": {
            "title": "Turn Count",
            "type": "integer"
          }
        },
        "required": [
          "conversation_id",
          "session_id",
          "turn_count",
          "started_at",
          "ended_at",
          "last_activity",
          "model",
          "effort_level",
          "config_version",
          "git_branch",
          "total_tokens",
          "total_tool_calls",
          "input_preview",
          "signal_evidence"
        ],
        "title": "SessionSummaryResponse",
        "type": "object"
      },
...
      "TrendEntryResponse": {
        "additionalProperties": false,
        "properties": {
          "delta": {
            "title": "Delta",
            "type": "number"
          },
          "direction": {
            "enum": [
              "regression",
              "improvement"
            ],
            "title": "Direction",
            "type": "string"
          },
          "older_mean": {
            "title": "Older Mean",
            "type": "number"
          },
          "recent_mean": {
            "title": "Recent Mean",
            "type": "number"
          },
          "sample_count": {
            "title": "Sample Count",
            "type": "integer"
          },
          "scorer": {
            "title": "Scorer",
            "type": "string"
          },
          "significant": {
            "title": "Significant",
            "type": "boolean"
          }
        },
        "required": [
          "scorer",
          "direction",
          "older_mean",
          "recent_mean",
          "delta",
          "sample_count",
          "significant"
        ],
        "title": "TrendEntryResponse",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "weave-agent-signals",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/api/analyze": {
      "get": {
        "operationId": "get_analysis_api_analyze_get",
        "parameters": [
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 1000,
              "maximum": 5000,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AnalysisResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Analysis",
        "tags": [
          "inspection"
        ]
      }
    },
...
    "/api/sessions": {
      "get": {
        "operationId": "get_sessions_api_sessions_get",
        "parameters": [
          {
            "in": "query",
            "name": "since",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Since"
            }
          },
          {
            "in": "query",
            "name": "until",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Until"
            }
          },
          {
            "in": "query",
            "name": "timezone",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Timezone"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 20,
              "maximum": 100,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionListResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Sessions",
        "tags": [
          "inspection"
        ]
      }
    },
    "/api/sessions/{conversation_id}": {
      "get": {
        "operationId": "get_session_api_sessions__conversation_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "conversation_id",
            "required": true,
            "schema": {
              "title": "Conversation Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionDetailResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Session",
        "tags": [
          "inspection"
        ]
      }
    }
  }
}